Added most existent files
These files in this upload pre-date this repository. Most of the work here was done thanks to Derpy Horse, as this stuff is mostly ship and outfit stats, though I must not neglect to thank local god and Star-Cross (Takarada Kaneo) for their work with the mod so far as well. Note that the file "antfolk map.txt" is not ready for testing, yet.